+++ packages/integration-jira/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @airnauts/airside-extension-jira
 
+## 0.10.1
+
+### Patch Changes
+
+- Updated dependencies [29c37c2]
+  - @airnauts/airside-server@0.10.1
+  - @airnauts/airside-core@0.10.1
+
+## 0.10.0
+
+### Patch Changes
+
+- Updated dependencies [2535eb8]
+  - @airnauts/airside-server@0.10.0
+  - @airnauts/airside-core@0.10.0
+
 ## 0.9.1
 
 ### Patch Changes